FixInterlacedFades: universal "match field closer to target color" mode
The target color may be something other than pure black or white,
such that pixel brightness may not only consistently fall
or consistently rise during the fade.

Even when a fade does have monotonically changing average brightness,
near the target end of the fade, compression noise may cause a field's
average to overshoot the target. In this case, current Darken/Brighten
modes treat this overshot field as the source of truth and modify
the other field, even if in reality it exactly matches the target color.

Furthermore, some pixels may be darker and some brighter than the target:
in the extreme case, if a bright-and-dark checkerboard pattern
fades to gray, the average brightness of the entire frame stays constant.
The individual pixels' absolute differences from the target, however,
are larger on one end of the fade and approach zero on the other end.

Finally, even on black/white targets, the target chroma is a mid-range
value (float 0 on a scale from -0.5 to +0.5). Current Darken handles this
by keeping the field whose absolute chroma is closer to zero
and adjusting the other field, which is correct, but current Brighten
does the opposite, moving the more-neutral chroma *away* from zero, thus
amplifying random color fluctuations near the target end of the fade.

This can all be fixed at once by taking the individual pixels' absolute
differences from the target color and simply adjusting the values towards
the target color, no matter what color it is. This additionally removes
any need for a human to decide whether to "darken" or "brighten" the fade,
thus making FixInterlacedFades somewhat easier to use.

When the target color is pure black or pure white, the new behaviour
for luma is the same as the current Darken/Brighten mode respectively.
For chroma, as well as for all other target colors,
it hopefully provides a strict improvement.

Fixes: https://github.com/Jaded-Encoding-Thaumaturgy/vs-deinterlace/issues/41
Fixes: https://github.com/Jaded-Encoding-Thaumaturgy/vs-deinterlace/issues/34