PackageStorage: fix folder setup when script path is used
Regression by d72222b. `cwd` may point to a file (if coming from the
`get_script_path()` branch) or a directory. There used to be a
`cwd = cwd.get_folder()` here to normalize it to a directory, but it was
removed. What happens then is that `cwd` may be
e.g. `/path/to/encode/01.py`, and then the constructed folder path ends
up being `/path/to/encode/01.py/.vsjet`, which won't work.